ci: convert to javascript
