+++ .ci/aui/buildContainerImage.js
@@ -1,4 +1,4 @@
-#!/usr/bin/env -S npx ts-node --esm -T
+#!/usr/bin/env node
 
 import { $ } from "zx";
 import { getRequiredEnvVar } from "../lib.js";
