+++ .ci/aui/pushContainerImage.js
@@ -1,8 +1,8 @@
-#!/usr/bin/env -S npx ts-node --esm
+#!/usr/bin/env node
 
 import { $ } from "zx";
 import { getRequiredEnvVar } from "../lib.js";
 
 const tag = getRequiredEnvVar("TAG");
 
-await $`docker push ghcr.io/nmshd/backbone-consumer-api:${tag}`;
+await $`docker push ghcr.io/nmshd/backbone-admin-ui:${tag}`;
