+++ .ci/capi/pushContainerImage.js
@@ -1,0 +1,8 @@
+#!/usr/bin/env node
+
+import { $ } from "zx";
+import { getRequiredEnvVar } from "../lib.js";
+
+const tag = getRequiredEnvVar("TAG");
+
+await $`docker push ghcr.io/nmshd/backbone-consumer-api:${tag}`;
