+++ .ci/helm/buildChart.js
@@ -1,8 +1,9 @@
-#!/usr/bin/env -S npx ts-node --esm -T
+#!/usr/bin/env node
 
 import { $ } from "zx";
 import { getRequiredEnvVar } from "../lib.js";
 
 const version = getRequiredEnvVar("VERSION");
 
-await $`helm push backbone-helm-chart-${version}.tgz oci://ghcr.io/nmshd`;
+await $`helm dependency update helm`;
+await $`helm package --version ${version} helm`;
