+++ .ci/test.js
@@ -1,4 +1,4 @@
-#!/usr/bin/env -S npx ts-node --esm
+#!/usr/bin/env node
 
 import { $ } from "zx";
